+++ apps/demo/app/[...puckPath]/_components/Handler.tsx
@@ -3,8 +3,6 @@ import React from "react";
 import { useEffect } from "react";
 
 export const Handler = () => {
-  const { updateOverride } = usePuck();
-
   // useEffect(() => {
   //   updateOverride(
   //     EOverrideFunctionId.onDragEnd,