+++ recipes/next/app/puck/[...puckPath]/page.tsx
@@ -11,7 +11,7 @@
  * NB this route is public, and you will need to add authentication
  */
 
-import "@measured/puck/puck.css";
+import "@helium-builder/puck/puck.css";
 import { Client } from "./client";
 import { Metadata } from "next";
 import { getPage } from "../../../lib/get-page";